Remove contentlayer from AI docs pipeline
