+++ apps/docs/app/ui/components/overview/util.ai.ts
@@ -1,0 +1,6 @@
+import { allComponents } from "@/app/lib/componentsFrontmatter.ai";
+import { getCategories } from "./categories.ts";
+
+export const categories = getCategories(allComponents);
+
+export { allComponents } from "@/app/lib/componentsFrontmatter.ai";
